:art: (jaewon/feature/back-end/#53) Post 관련 통신을 위한 파일들 구조 개선
