+++ Plinic/Back-End/PlinicAPI/PlinicAPI.swift
@@ -1,0 +1,14 @@
+//
+//  PlinicAPI.swift
+//  Plinic
+//
+//  Created by Jaewon on 2022/10/22.
+//
+
+import Foundation
+
+final class PlinicAPI {
+    
+    
+    
+}
