+++ Plinic/Back-End/PostAPI/DTO/Author.swift
@@ -1,0 +1,19 @@
+//
+//  Author.swift
+//  Plinic
+//
+//  Created by Jaewon on 2022/10/22.
+//
+
+import Foundation
+
+struct Author: Codable {
+    
+    let nickname: String
+    var profilePic: String?
+    
+    enum CodingKeys: String, CodingKey {
+        case nickname
+        case profilePic = "profile_pic"
+    }
+}
